Collection prototype

This prototype represents a collection of ProjectModifier objects.

Fields

$...

ProjectModifier objects stored in the collection.

Usage

x$print()

x$show()

x$repr()

x$ids()

x$length()

x$add

x$remove(id)

x$get_parameter(id)

x$set_parameter(id, value)

x$render_parameter(id)

x$render_all_parameters()

Arguments

id

id object.

value

any object.

Details

print

print the object.

show

show the object.

repr

character representation of object.

ids

character ids for objects inside collection.

length

integer number of objects inside collection.

add

add ProjectModifier object.

remove

remove an item from the collection.

get_parameter

retrieve the value of a parameter in the object using an id object.

set_parameter

change the value of a parameter in the object to a new object.

render_parameter

generate a shiny widget to modify the the value of a parameter (specified by argument id).

render_all_parameters

generate a shiny::div() containing all the parameters" widgets.

See also