Constraint class

This class is used to represent the constraints used in optimization. Only experts should use the fields and methods for this class directly.

See also

Super class

oppr::ProjectModifier -> Constraint

Methods

Public methods

Constraint$clone()

Inherited methods

oppr::ProjectModifier$apply()

oppr::ProjectModifier$calculate()

oppr::ProjectModifier$get_data()

oppr::ProjectModifier$get_internal()

oppr::ProjectModifier$print()

oppr::ProjectModifier$repr()

oppr::ProjectModifier$set_data()

oppr::ProjectModifier$set_internal()

oppr::ProjectModifier$show()

Method clone()

The objects of this class are cloneable with this method.

Usage

Constraint$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.