Multi-objective project problem class

This class is used to represent multi-objective project planning problems. It stores the data (e.g., actions, and features) and mathematical formulation (e.g., the objective, constraints, and other design criteria) needed to generate prioritizations. Most users should use multi_problem() to generate new multi-objective project problem objects, and the functions distributed with the package to interact with them. Only experts should use the fields and methods for this class directly.

See also

Public fields

problems

list containing ProjectProblem objects.

defaults

list indicating if other fields contain defaults.

approach

MultiObjApproach object for specifying the multi-objective optimization approach.

solver

Solver object specifying the solver for generating solutions.

Methods

Public methods

MultiObjProjectProblem$new()

MultiObjProjectProblem$print()

MultiObjProjectProblem$show()

MultiObjProjectProblem$repr()

MultiObjProjectProblem$number_of_problems()

MultiObjProjectProblem$number_of_features()

MultiObjProjectProblem$number_of_actions()

MultiObjProjectProblem$number_of_projects()

MultiObjProjectProblem$problem_names()

MultiObjProjectProblem$feature_names()

MultiObjProjectProblem$action_names()

MultiObjProjectProblem$project_names()

MultiObjProjectProblem$add_approach()

MultiObjProjectProblem$add_solver()

MultiObjProjectProblem$clone()

Method new()

Create a new multi-objective conservation problem object.

Usage

MultiObjProjectProblem$new(problems)

Arguments

problems

list containing ProjectProblem objects.

Returns

A new MultiObjProjectProblem object.

Method print()

Print concise information about the object.

Usage

MultiObjProjectProblem$print()

Returns

Invisible TRUE.

Method show()

Display concise information about the object.

Usage

MultiObjProjectProblem$show()

Returns

Invisible TRUE.

Method repr()

Generate a character representation of the object.

Usage

MultiObjProjectProblem$repr()

Returns

A character value.

Method number_of_problems()

Obtain the number of problems.

Usage

MultiObjProjectProblem$number_of_problems()

Returns

An integer value.

Method number_of_features()

Obtain the number of features.

Usage

MultiObjProjectProblem$number_of_features()

Returns

An integer value.

Method number_of_actions()

Obtain the number of actions.

Usage

MultiObjProjectProblem$number_of_actions()

Returns

An integer value.

Method number_of_projects()

Obtain the number of projects.

Usage

MultiObjProjectProblem$number_of_projects()

Returns

An integer value.

Method problem_names()

Obtain the names of the problems.

Usage

MultiObjProjectProblem$problem_names()

Returns

An character value.

Method feature_names()

Obtain the names of the features.

Usage

MultiObjProjectProblem$feature_names()

Returns

A list of character vectors.

Method action_names()

Obtain the names of the actions.

Usage

MultiObjProjectProblem$action_names()

Returns

A character vector.

Method project_names()

Obtain the names of the projects.

Usage

MultiObjProjectProblem$project_names()

Returns

A list of character vectors.

Method add_approach()

Create a new object with an approach added to the problem formulation.

Usage

MultiObjProjectProblem$add_approach(x)

Arguments

x

MultiObjApproach object.

Returns

An updated MultiObjProjectProblem object.

Method add_solver()

Create a new object with a solver added to the problem formulation.

Usage

MultiObjProjectProblem$add_solver(x)

Arguments

x

Solver object.

Returns

An updated MultiObjProjectProblem object.

Method clone()

The objects of this class are cloneable with this method.

Usage

MultiObjProjectProblem$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.