Parameter class

This class is used to represent a parameter that has multiple values. Each value has a different label to differentiate values. Only experts should interact directly with this class.

Fields

$id

Id() identifier for parameter.

$name

character name of parameter.

$value

numeric vector of values.

$default

numeric vector of default values.

$class

character name of the class that the values inherit from (e.g. "integer".

$lower_limit

numeric vector specifying the minimum permitted value for each element in $value.

$upper_limit

numeric vector specifying the maximum permitted value for each element in $value.

$widget

function used to construct a shiny::shiny() interface for modifying values.

Usage

x$print()

x$show()

x$reset()

Details

print

print the object.

show

show the object.

reset

change the parameter values to be the default values.

See also