Parameters class

This class represents a collection of Parameter objects. It provides methods for accessing, updating, and rendering the parameters stored inside it.

Fields

$parameters

list object containing Parameter objects.

Usage

x$print()

x$show()

x$repr()

x$names()

x$ids()

x$length()

x$get(id)

x$set(id, value)

x$add(p)

x$render(id)

x$render_all()

Arguments

id

Id() object.

p

Parameter object.

value

any object.

Details

print

print the object.

show

show the object.

repr

character representation of object.

names

return character names of parameters.

ids

return character parameter unique identifiers.

length

return integer number of parameters in object.

get

retrieve the value of a parameter in the object using an Id object.

set

change the value of a parameter in the object to a new object.

render

generate a shiny widget to modify the the value of a parameter (specified by argument Id).

render_all

generate a shiny::div() containing all the parameters" widgets.