Conservation problem modifier class

Conservation problem modifier class

Conservation problem modifier class

Details

This super-class is used to represent prototypes that in turn are used to modify a ProjectProblem object. Specifically, the Constraint, Decision, Objective, and Target prototypes inherit from this class. Only experts should use the fields and methods for this class directly.

See also

Public fields

name

character value.

data

list containing data.

internal

list containing internal computed values.

Methods

Public methods

ProjectModifier$print()

ProjectModifier$show()

ProjectModifier$repr()

ProjectModifier$get_data()

ProjectModifier$set_data()

ProjectModifier$get_internal()

ProjectModifier$set_internal()

ProjectModifier$calculate()

ProjectModifier$apply()

ProjectModifier$clone()

Method print()

Print information about the object.

Usage

ProjectModifier$print()

Returns

None.

Method show()

Print information about the object.

Usage

ProjectModifier$show()

Returns

None.

Method repr()

Generate a character representation of the object.

Usage

ProjectModifier$repr()

Returns

A character value.

Method get_data()

Get values stored in the data field.

Usage

ProjectModifier$get_data(x)

Arguments

x

character name of data.

Returns

An object. If the data field does not contain an object associated with the argument to x, then a new_waiver() object is returned. Set values stored in the data field. Note that this method will overwrite existing data.

Method set_data()

Usage

ProjectModifier$set_data(x, value)

Arguments

x

character name of data.

value

Object to store.

Returns

Invisible TRUE.

Method get_internal()

Get values stored in the internal field.

Usage

ProjectModifier$get_internal(x)

Arguments

x

character name of data.

Returns

An object. If the internal field does not contain an object associated with the argument to x, then a new_waiver() object is returned.

Method set_internal()

Set values stored in the internal field. Note that this method will overwrite existing data.

Usage

ProjectModifier$set_internal(x, value)

Arguments

x

character name of data.

value

Object to store.

Returns

An object. If the internal field does not contain an object associated with the argument to x, then a new_waiver() object is returned.

Method calculate()

Perform computations that need to be completed before applying the object.

Usage

ProjectModifier$calculate(y)

Arguments

y

problem() object.

x

new_optimization_problem() object.

Returns

Invisible TRUE.

Method apply()

Update an optimization problem formulation.

Usage

ProjectModifier$apply(x)

Arguments

x

new_optimization_problem() object.

Returns

Invisible TRUE.

Method clone()

The objects of this class are cloneable with this method.

Usage

ProjectModifier$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.