Project problem class

Project problem class

Project problem class

Description

This class is used to represent project prioritization problems. Only experts should use the fields and methods for this class directly.

See also

Public fields

data

list containing data (e.g., projects, features).

defaults

list indicating if other fields contain defaults.

objective

Objective object used to specify the objective function.

weights

Weight object used to specify the objective function.

decisions

Decision object used to represent the type of decision made on planning units.

targets

Target object used to represent representation targets for features.

constraints

list object used to store Constraint objects for the problem.

solver

Solver object used to specify the process for generating a solution.

Methods

Public methods

ProjectProblem$new()

ProjectProblem$print()

ProjectProblem$show()

ProjectProblem$repr()

ProjectProblem$get_data()

ProjectProblem$set_data()

ProjectProblem$number_of_actions()

ProjectProblem$number_of_projects()

ProjectProblem$number_of_features()

ProjectProblem$action_names()

ProjectProblem$project_names()

ProjectProblem$feature_names()

ProjectProblem$feature_weights()

ProjectProblem$feature_targets()

ProjectProblem$feature_phylogeny()

ProjectProblem$action_costs()

ProjectProblem$project_costs()

ProjectProblem$project_success_probabilities()

ProjectProblem$of_matrix()

ProjectProblem$pa_matrix()

ProjectProblem$eof_matrix()

ProjectProblem$add_solver()

ProjectProblem$add_targets()

ProjectProblem$add_weights()

ProjectProblem$add_objective()

ProjectProblem$add_decisions()

ProjectProblem$add_constraint()

ProjectProblem$clone()

Method new()

Create a new conservation problem object.

Usage

ProjectProblem$new(data = list())

Arguments

data

list containing data

Returns

A new ProjectProblem object.

Method print()

Print concise information about the object.

Usage

ProjectProblem$print()

Returns

Invisible TRUE.

Method show()

Display concise information about the object.

Usage

ProjectProblem$show()

Returns

Invisible TRUE.

Method repr()

Generate a character representation of the object.

Usage

ProjectProblem$repr()

Returns

A character value.

Method get_data()

Get values stored in the data field.

Usage

ProjectProblem$get_data(x)

Arguments

x

character name of data.

Returns

An object. If the data field does not contain an object associated with the argument to x, then a new_waiver() object is returned.

Method set_data()

Set values stored in the data field. Note that this method will overwrite existing data.

Usage

ProjectProblem$set_data(x, value)

Arguments

x

character name of data.

value

Object to store.

Returns

Invisible TRUE.

Method number_of_actions()

Obtain the number of actions.

Usage

ProjectProblem$number_of_actions()

Returns

An integer value.

Method number_of_projects()

Obtain the number of projects.

Usage

ProjectProblem$number_of_projects()

Returns

An integer value.

Method number_of_features()

Obtain the number of features.

Usage

ProjectProblem$number_of_features()

Returns

An integer value.

Method action_names()

Obtain the names of the actions.

Usage

ProjectProblem$action_names()

Returns

A character vector.

Method project_names()

Obtain the names of the projects.

Usage

ProjectProblem$project_names()

Returns

A character vector.

Method feature_names()

Obtain the names of the features.

Usage

ProjectProblem$feature_names()

Returns

A character vector.

Method feature_weights()

Obtain the feature weights.

Usage

ProjectProblem$feature_weights()

Returns

A named numeric vector.

Method feature_targets()

Obtain the feature targets.

Usage

ProjectProblem$feature_targets()

Returns

A tibble::tibble() object.

Method feature_phylogeny()

Obtain the feature phylogeny.

Usage

ProjectProblem$feature_phylogeny()

Returns

A ape::phylo() phylogenetic tree object.

Method action_costs()

Obtain the action costs.

Usage

ProjectProblem$action_costs()

Returns

A numeric vector.

Method project_costs()

Obtain the project costs.

Usage

ProjectProblem$project_costs()

Returns

A numeric vector.

Method project_success_probabilities()

Obtain the probability that each project will succeed if funded.

Usage

ProjectProblem$project_success_probabilities()

Returns

A numeric vector.

Method of_matrix()

Obtain information on the outcome for each feature that would be expected if each project funded and is successfully completed.

Usage

ProjectProblem$of_matrix()

Returns

A Matrix::dgCMatrix object.

Method pa_matrix()

Obtain information on which actions are associated with each project.

Usage

ProjectProblem$pa_matrix()

Returns

A Matrix::dgCMatrix object.

Method eof_matrix()

Calculate the expected outcome for each feature assuming that each project is funded and accounting for the possibility that funded projects may fail to be successfully completed.

Usage

ProjectProblem$eof_matrix()

Returns

A Matrix::dgCMatrix object.

Method add_solver()

Create a new object with a solver added to the problem formulation.

Usage

ProjectProblem$add_solver(x)

Arguments

x

Solver object.

Returns

An updated ProjectProblem object.

Method add_targets()

Create a new object with targets added to the problem formulation.

Usage

ProjectProblem$add_targets(x)

Arguments

x

Target object.

Returns

An updated ProjectProblem object.

Method add_weights()

Create a new object with weights added to the problem formulation.

Usage

ProjectProblem$add_weights(x)

Arguments

x

Weight object.

Returns

An updated ProjectProblem object.

Method add_objective()

Create a new object with the objective added to the problem formulation.

Usage

ProjectProblem$add_objective(x)

Arguments

x

Objective object.

Returns

An updated ProjectProblem object.

Method add_decisions()

Create a new object with the decisions added to the problem formulation.

Usage

ProjectProblem$add_decisions(x)

Arguments

x

Objective object.

Returns

An updated ProjectProblem object.

Method add_constraint()

Create a new object with the constraint added to the problem formulation.

Usage

ProjectProblem$add_constraint(x)

Arguments

x

Constraint object.

Returns

An updated ProjectProblem object.

Method clone()

The objects of this class are cloneable with this method.

Usage

ProjectProblem$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.