Solver class

This class is used to represent solvers for optimization. Only experts should use the fields and methods for this class directly.

See also

Super class

ProjectModifier -> Solver

Public fields

has_pwlobj

logical indicating if solver supports piece-wise linear components in an objective function.

Methods

Public methods

Solver$set_start_solution()

Solver$remove_start_solution()

Solver$solve()

Solver$clone()

Inherited methods

ProjectModifier$apply()

ProjectModifier$calculate()

ProjectModifier$get_data()

ProjectModifier$get_internal()

ProjectModifier$print()

ProjectModifier$repr()

ProjectModifier$set_data()

ProjectModifier$set_internal()

ProjectModifier$show()

Solver$set_start_solution()

Set start solution.

Usage

Solver$set_start_solution(x)

Arguments

x

numeric vector.

Returns

Invisible TRUE.

Solver$remove_start_solution()

Remove start solution.

Usage

Solver$remove_start_solution(x)

Arguments

x

numeric vector.

Returns

Invisible TRUE.

Solver$solve()

Solve an optimization problem.

Usage

Solver$solve(x, ...)

Arguments

x

new_optimization_problem() object.

...

Additional arguments as needed.

Returns

Invisible TRUE.

Solver$clone()

The objects of this class are cloneable with this method.

Usage

Solver$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.