Solver prototype

This prototype is used to generate objects that represent methods for solving optimization problems. This class represents a recipe to create solver and and is only recommended for use by expert users. To customize the method used to solve optimization problems, please see the help page on solvers.

Fields

$name

character name of solver.

$parameters

Parameters object with parameters used to customize the the solver.

$solve

function used to solve a OptimizationProblem object.

Usage

x$print()

x$show()

x$repr()

x$solve(op)

Arguments

x

Solver object.

op

OptimizationProblem object.

Details

print

print the object.

show

show the object.

repr

character representation of object.

solve

solve an OptimizationProblem using this object.