Multi-objective optimization approaches

Approaches specify methods for generating solutions for multi-objective optimization problems.

Details

The following approaches can be used to generate solutions for a multi-objective project prioritization problem.

add_abs_constraint_approach()

Add an approach to generate solutions based on constraints that specify the required objectives values.

add_wtd_goal_approach()

Add an approach to generate solutions with the weighted goal method.

add_chebyshev_approach()

Add an approach to generate solutions with the Chebyshev method.

add_ref_point_approach()

Add an approach to generate solutions with the reference point method.

See also

Examples

# \dontrun{
# load data
data(sim_multi_projects)
data(sim_multi_features)
data(sim_multi_actions)
data(sim_multi_tree)

# build problem
p1 <-
  multi_problem(
    obj1 =
      problem(
        sim_multi_projects[[1]], sim_multi_actions, sim_multi_features[[1]],
        "name", "success", "name", "cost", "name",
        baseline_project_name = "baseline_project_obj1"
      ) %>%
      add_max_phylo_div_objective(
       budget = 1000, tree = sim_multi_tree[[1]]
      ) %>%
      add_binary_decisions(),
   obj2 =
     problem(
       sim_multi_projects[[2]], sim_multi_actions, sim_multi_features[[2]],
       "name", "success", "name", "cost", "name",
       baseline_project_name = "baseline_project_obj2"
     ) %>%
     add_max_richness_objective(budget = 1000) %>%
     add_binary_decisions(),
   obj3 =
     problem(
       sim_multi_projects[[3]], sim_multi_actions, sim_multi_features[[3]],
       "name", "success", "name", "cost", "name",
       baseline_project_name = "baseline_project_obj3"
     ) %>%
     add_max_wtd_sum_objective(budget = 1000) %>%
     add_binary_decisions()
 ) %>%
 add_default_solver()

# build another problem, with the absolute constraint method
p2 <-
  p1 %>%
  add_abs_constraint_approach(
    goals = c(NA, 0.01, 0.01)
  )

# build another problem, with the weighted goal method
p3 <-
  p1 %>%
  add_wtd_goal_approach(
    weights = c(1, 0.5, 0.1),
    goals = c(1, 3, 0.2)
  )

# build another problem, with the Chebyshev method
p4 <-
  p1 %>%
  add_chebyshev_approach(
    weights = c(1, 0.5, 0.1),
    goals = c(1, 3, 0.2)
  )

# build another problem, with the reference point method
p5 <-
  p1 %>%
  add_ref_point_approach(
    weights = c(1, 0.5, 0.1),
    goals = c(1, 3, 0.2)
  )

# generate solutions using each approach
s <- rbind(solve(p2), solve(p3), solve(p4), solve(p5))
#> Set parameter Username
#> Set parameter LicenseID to value 2738655
#> Set parameter TimeLimit to value 2147483647
#> Set parameter MIPGap to value 0
#> Set parameter NumericFocus to value 2
#> Set parameter Presolve to value 2
#> Set parameter Threads to value 1
#> Set parameter PoolSolutions to value 1
#> Set parameter PoolSearchMode to value 2
#> Academic license - for non-commercial use only - expires 2026-11-14
#> Gurobi Optimizer version 13.0.0 build v13.0.0rc1 (linux64 - "Ubuntu 24.04.2 LTS")
#> 
#> CPU model: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz, instruction set [SSE2|AVX|AVX2|AVX512]
#> Thread count: 4 physical cores, 8 logical processors, using up to 1 threads
#> 
#> Non-default parameters:
#> TimeLimit  2147483647
#> MIPGap  0
#> NumericFocus  2
#> Presolve  2
#> Threads  1
#> PoolSolutions  1
#> PoolSearchMode  2
#> 
#> Optimize a model with 334 rows, 260 columns and 918 nonzeros (Max)
#> Model fingerprint: 0x6755096a
#> Model has 201 linear objective coefficients
#> Variable types: 11 continuous, 150 integer (150 binary)
#> Semi-Variable types: 99 continuous, 0 integer
#> Coefficient statistics:
#>   Matrix range     [5e-02, 1e+02]
#>   Objective range  [2e-02, 5e-01]
#>   Bounds range     [6e-01, 3e+00]
#>   RHS range        [1e-02, 1e+03]
#> Presolve removed 244 rows and 80 columns
#> Presolve time: 0.00s
#> Presolved: 286 rows, 278 columns, 768 nonzeros
#> Variable types: 98 continuous, 180 integer (180 binary)
#> Found heuristic solution: objective 0.3873963
#> Root relaxation presolved: 286 rows, 278 columns, 768 nonzeros
#> 
#> 
#> Root relaxation: objective 8.326604e-01, 71 iterations, 0.00 seconds (0.00 work units)
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0    0.83266    0   13    0.38740    0.83266   115%     -    0s
#> H    0     0                       0.6823882    0.83266  22.0%     -    0s
#>      0     0    0.81807    0   16    0.68239    0.81807  19.9%     -    0s
#>      0     0    0.81807    0   12    0.68239    0.81807  19.9%     -    0s
#>      0     0    0.81807    0   15    0.68239    0.81807  19.9%     -    0s
#>      0     0    0.79417    0   19    0.68239    0.79417  16.4%     -    0s
#>      0     0    0.71938    0    3    0.68239    0.71938  5.42%     -    0s
#> H    0     0                       0.7096175    0.71938  1.38%     -    0s
#>      0     0    0.71936    0    5    0.70962    0.71936  1.37%     -    0s
#>      0     0    0.71936    0    3    0.70962    0.71936  1.37%     -    0s
#>      0     0    0.71936    0    5    0.70962    0.71936  1.37%     -    0s
#>      0     0    0.71936    0    5    0.70962    0.71936  1.37%     -    0s
#>      0     0    0.71856    0    7    0.70962    0.71856  1.26%     -    0s
#> H    0     0                       0.7113862    0.71849  1.00%     -    0s
#>      0     0    0.71849    0    7    0.71139    0.71849  1.00%     -    0s
#>      0     0    0.71849    0    7    0.71139    0.71849  1.00%     -    0s
#>      0     2    0.71847    0    7    0.71139    0.71847  1.00%     -    0s
#> 
#> Cutting planes:
#>   Cover: 17
#>   MIR: 1
#>   Flow cover: 1
#> 
#> Explored 167 nodes (630 simplex iterations) in 0.05 seconds (0.03 work units)
#> Thread count was 1 (of 8 available processors)
#> 
#> Solution count 1: 0.711386 
#> No other solutions better than 0.711386
#> 
#> Optimal solution found (tolerance 0.00e+00)
#> Best objective 7.113861930761e-01, best bound 7.113861930761e-01, gap 0.0000%
#> Set parameter Username
#> Set parameter LicenseID to value 2738655
#> Set parameter TimeLimit to value 2147483647
#> Set parameter MIPGap to value 0
#> Set parameter NumericFocus to value 2
#> Set parameter Presolve to value 2
#> Set parameter Threads to value 1
#> Set parameter PoolSolutions to value 1
#> Set parameter PoolSearchMode to value 2
#> Academic license - for non-commercial use only - expires 2026-11-14
#> Gurobi Optimizer version 13.0.0 build v13.0.0rc1 (linux64 - "Ubuntu 24.04.2 LTS")
#> 
#> CPU model: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz, instruction set [SSE2|AVX|AVX2|AVX512]
#> Thread count: 4 physical cores, 8 logical processors, using up to 1 threads
#> 
#> Non-default parameters:
#> TimeLimit  2147483647
#> MIPGap  0
#> NumericFocus  2
#> Presolve  2
#> Threads  1
#> PoolSolutions  1
#> PoolSearchMode  2
#> 
#> Optimize a model with 335 rows, 263 columns and 1122 nonzeros (Min)
#> Model fingerprint: 0x712f3a04
#> Model has 3 linear objective coefficients
#> Variable types: 14 continuous, 150 integer (150 binary)
#> Semi-Variable types: 99 continuous, 0 integer
#> Coefficient statistics:
#>   Matrix range     [2e-02, 1e+02]
#>   Objective range  [1e-01, 1e+00]
#>   Bounds range     [6e-01, 3e+00]
#>   RHS range        [2e-01, 1e+03]
#> Presolve removed 243 rows and 80 columns
#> Presolve time: 0.01s
#> Presolved: 288 rows, 281 columns, 924 nonzeros
#> Variable types: 101 continuous, 180 integer (180 binary)
#> Found heuristic solution: objective 1.0409811
#> Found heuristic solution: objective 0.9421745
#> Root relaxation presolved: 288 rows, 281 columns, 924 nonzeros
#> 
#> 
#> Root relaxation: objective 5.592195e-01, 160 iterations, 0.00 seconds (0.00 work units)
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0    0.55922    0   16    0.94217    0.55922  40.6%     -    0s
#> H    0     0                       0.7459892    0.55922  25.0%     -    0s
#>      0     0    0.60010    0   17    0.74599    0.60010  19.6%     -    0s
#>      0     0    0.62724    0   14    0.74599    0.62724  15.9%     -    0s
#> H    0     0                       0.7413913    0.62724  15.4%     -    0s
#>      0     0    0.62724    0   16    0.74139    0.62724  15.4%     -    0s
#>      0     0    0.62724    0   16    0.74139    0.62724  15.4%     -    0s
#>      0     0    0.62724    0   12    0.74139    0.62724  15.4%     -    0s
#>      0     0    0.62724    0   15    0.74139    0.62724  15.4%     -    0s
#>      0     0    0.65561    0   10    0.74139    0.65561  11.6%     -    0s
#>      0     0    0.70891    0    3    0.74139    0.70891  4.38%     -    0s
#>      0     0    0.70899    0    3    0.74139    0.70899  4.37%     -    0s
#>      0     0    0.70901    0    5    0.74139    0.70901  4.37%     -    0s
#>      0     0    0.70911    0    7    0.74139    0.70911  4.35%     -    0s
#>      0     0    0.70911    0    7    0.74139    0.70911  4.35%     -    0s
#> H    0     0                       0.7392152    0.70911  4.07%     -    0s
#>      0     0    0.70912    0    3    0.73922    0.70912  4.07%     -    0s
#>      0     0    0.70912    0    5    0.73922    0.70912  4.07%     -    0s
#>      0     0    0.70912    0    7    0.73922    0.70912  4.07%     -    0s
#>      0     0    0.70985    0    6    0.73922    0.70985  3.97%     -    0s
#> H    0     0                       0.7187599    0.70987  1.24%     -    0s
#>      0     0    0.70987    0    8    0.71876    0.70987  1.24%     -    0s
#>      0     0    0.70987    0   11    0.71876    0.70987  1.24%     -    0s
#>      0     0    0.70987    0   11    0.71876    0.70987  1.24%     -    0s
#>      0     0    0.70987    0   11    0.71876    0.70987  1.24%     -    0s
#> H    0     0                       0.7169912    0.70990  0.99%     -    0s
#>      0     2    0.70990    0   11    0.71699    0.70990  0.99%     -    0s
#> 
#> Cutting planes:
#>   Cover: 15
#>   MIR: 4
#>   Flow cover: 1
#>   RLT: 1
#> 
#> Explored 174 nodes (815 simplex iterations) in 0.09 seconds (0.05 work units)
#> Thread count was 1 (of 8 available processors)
#> 
#> Solution count 1: 0.716991 
#> No other solutions better than 0.716991
#> 
#> Optimal solution found (tolerance 0.00e+00)
#> Best objective 7.169911970196e-01, best bound 7.169911970196e-01, gap 0.0000%
#> Set parameter Username
#> Set parameter LicenseID to value 2738655
#> Set parameter TimeLimit to value 2147483647
#> Set parameter MIPGap to value 0
#> Set parameter NumericFocus to value 2
#> Set parameter Presolve to value 2
#> Set parameter Threads to value 1
#> Set parameter PoolSolutions to value 1
#> Set parameter PoolSearchMode to value 2
#> Academic license - for non-commercial use only - expires 2026-11-14
#> Gurobi Optimizer version 13.0.0 build v13.0.0rc1 (linux64 - "Ubuntu 24.04.2 LTS")
#> 
#> CPU model: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz, instruction set [SSE2|AVX|AVX2|AVX512]
#> Thread count: 4 physical cores, 8 logical processors, using up to 1 threads
#> 
#> Non-default parameters:
#> TimeLimit  2147483647
#> MIPGap  0
#> NumericFocus  2
#> Presolve  2
#> Threads  1
#> PoolSolutions  1
#> PoolSearchMode  2
#> 
#> Optimize a model with 338 rows, 264 columns and 1128 nonzeros (Min)
#> Model fingerprint: 0xbf6cf508
#> Model has 1 linear objective coefficients
#> Variable types: 15 continuous, 150 integer (150 binary)
#> Semi-Variable types: 99 continuous, 0 integer
#> Coefficient statistics:
#>   Matrix range     [2e-02, 1e+02]
#>   Objective range  [1e+00, 1e+00]
#>   Bounds range     [6e-01, 3e+00]
#>   RHS range        [2e-01, 1e+03]
#> Presolve removed 244 rows and 80 columns
#> Presolve time: 0.01s
#> Presolved: 290 rows, 282 columns, 928 nonzeros
#> Variable types: 102 continuous, 180 integer (180 binary)
#> Found heuristic solution: objective 0.6126037
#> Root relaxation presolved: 290 rows, 282 columns, 928 nonzeros
#> 
#> 
#> Root relaxation: objective 3.408968e-01, 130 iterations, 0.00 seconds (0.00 work units)
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0    0.34090    0   23    0.61260    0.34090  44.4%     -    0s
#>      0     0    0.34844    0   25    0.61260    0.34844  43.1%     -    0s
#> H    0     0                       0.4283774    0.42292  1.27%     -    0s
#>      0     0    0.42292    0   26    0.42838    0.42292  1.27%     -    0s
#>      0     0    0.42292    0   21    0.42838    0.42292  1.27%     -    0s
#>      0     0    0.42292    0   14    0.42838    0.42292  1.27%     -    0s
#>      0     0    0.42292    0   20    0.42838    0.42292  1.27%     -    0s
#>      0     0    0.42292    0   15    0.42838    0.42292  1.27%     -    0s
#>      0     0     cutoff    0         0.42838    0.42838  0.00%     -    0s
#> 
#> Explored 1 nodes (282 simplex iterations) in 0.03 seconds (0.03 work units)
#> Thread count was 1 (of 8 available processors)
#> 
#> Solution count 1: 0.428377 
#> No other solutions better than 0.428377
#> 
#> Optimal solution found (tolerance 0.00e+00)
#> Best objective 4.283773900957e-01, best bound 4.283773900957e-01, gap 0.0000%
#> Set parameter Username
#> Set parameter LicenseID to value 2738655
#> Set parameter TimeLimit to value 2147483647
#> Set parameter MIPGap to value 0
#> Set parameter NumericFocus to value 2
#> Set parameter Presolve to value 2
#> Set parameter Threads to value 1
#> Set parameter PoolSolutions to value 1
#> Set parameter PoolSearchMode to value 2
#> Academic license - for non-commercial use only - expires 2026-11-14
#> Gurobi Optimizer version 13.0.0 build v13.0.0rc1 (linux64 - "Ubuntu 24.04.2 LTS")
#> 
#> CPU model: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz, instruction set [SSE2|AVX|AVX2|AVX512]
#> Thread count: 4 physical cores, 8 logical processors, using up to 1 threads
#> 
#> Non-default parameters:
#> TimeLimit  2147483647
#> MIPGap  0
#> NumericFocus  2
#> Presolve  2
#> Threads  1
#> PoolSolutions  1
#> PoolSearchMode  2
#> 
#> Optimize a model with 338 rows, 264 columns and 1128 nonzeros (Min)
#> Model fingerprint: 0xbf6cf508
#> Model has 1 linear objective coefficients
#> Variable types: 15 continuous, 150 integer (150 binary)
#> Semi-Variable types: 99 continuous, 0 integer
#> Coefficient statistics:
#>   Matrix range     [2e-02, 1e+02]
#>   Objective range  [1e+00, 1e+00]
#>   Bounds range     [6e-01, 3e+00]
#>   RHS range        [2e-01, 1e+03]
#> Presolve removed 244 rows and 80 columns
#> Presolve time: 0.01s
#> Presolved: 290 rows, 282 columns, 928 nonzeros
#> Variable types: 102 continuous, 180 integer (180 binary)
#> Found heuristic solution: objective 0.6126037
#> Root relaxation presolved: 290 rows, 282 columns, 928 nonzeros
#> 
#> 
#> Root relaxation: objective 3.408968e-01, 130 iterations, 0.00 seconds (0.00 work units)
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0    0.34090    0   23    0.61260    0.34090  44.4%     -    0s
#>      0     0    0.34844    0   25    0.61260    0.34844  43.1%     -    0s
#> H    0     0                       0.4283774    0.42292  1.27%     -    0s
#>      0     0    0.42292    0   26    0.42838    0.42292  1.27%     -    0s
#>      0     0    0.42292    0   21    0.42838    0.42292  1.27%     -    0s
#>      0     0    0.42292    0   14    0.42838    0.42292  1.27%     -    0s
#>      0     0    0.42292    0   20    0.42838    0.42292  1.27%     -    0s
#>      0     0    0.42292    0   15    0.42838    0.42292  1.27%     -    0s
#>      0     0     cutoff    0         0.42838    0.42838  0.00%     -    0s
#> 
#> Explored 1 nodes (282 simplex iterations) in 0.03 seconds (0.03 work units)
#> Thread count was 1 (of 8 available processors)
#> 
#> Solution count 1: 0.428377 
#> No other solutions better than 0.428377
#> 
#> Optimal solution found (tolerance 0.00e+00)
#> Best objective 4.283773900957e-01, best bound 4.283773900957e-01, gap 0.0000%
#> Set parameter Username
#> Set parameter LicenseID to value 2738655
#> Set parameter TimeLimit to value 2147483647
#> Set parameter MIPGap to value 0
#> Set parameter NumericFocus to value 2
#> Set parameter Presolve to value 2
#> Set parameter Threads to value 1
#> Set parameter PoolSolutions to value 1
#> Set parameter PoolSearchMode to value 2
#> Academic license - for non-commercial use only - expires 2026-11-14
#> Gurobi Optimizer version 13.0.0 build v13.0.0rc1 (linux64 - "Ubuntu 24.04.2 LTS")
#> 
#> CPU model: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz, instruction set [SSE2|AVX|AVX2|AVX512]
#> Thread count: 4 physical cores, 8 logical processors, using up to 1 threads
#> 
#> Non-default parameters:
#> TimeLimit  2147483647
#> MIPGap  0
#> NumericFocus  2
#> Presolve  2
#> Threads  1
#> PoolSolutions  1
#> PoolSearchMode  2
#> 
#> Optimize a model with 339 rows, 264 columns and 1129 nonzeros (Min)
#> Model fingerprint: 0x5d6761e9
#> Model has 3 linear objective coefficients
#> Variable types: 15 continuous, 150 integer (150 binary)
#> Semi-Variable types: 99 continuous, 0 integer
#> Coefficient statistics:
#>   Matrix range     [2e-02, 1e+02]
#>   Objective range  [1e-01, 1e+00]
#>   Bounds range     [6e-01, 3e+00]
#>   RHS range        [2e-01, 1e+03]
#> 
#> Loaded user MIP start with objective 0.856755
#> 
#> Presolve removed 245 rows and 80 columns
#> Presolve time: 0.01s
#> Presolved: 290 rows, 282 columns, 928 nonzeros
#> Variable types: 102 continuous, 180 integer (180 binary)
#> Root relaxation presolved: 290 rows, 282 columns, 928 nonzeros
#> 
#> 
#> Root relaxation: objective 5.592195e-01, 146 iterations, 0.00 seconds (0.00 work units)
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0    0.55922    0   16    0.85675    0.55922  34.7%     -    0s
#> H    0     0                       0.7459892    0.55922  25.0%     -    0s
#>      0     0    0.59805    0   16    0.74599    0.59805  19.8%     -    0s
#> H    0     0                       0.7413913    0.63860  13.9%     -    0s
#>      0     0    0.63860    0   17    0.74139    0.63860  13.9%     -    0s
#>      0     0    0.63860    0   15    0.74139    0.63860  13.9%     -    0s
#>      0     0    0.63860    0   15    0.74139    0.63860  13.9%     -    0s
#>      0     0    0.63860    0   16    0.74139    0.63860  13.9%     -    0s
#>      0     0    0.63860    0   22    0.74139    0.63860  13.9%     -    0s
#> H    0     0                       0.7371219    0.63860  13.4%     -    0s
#> H    0     0                       0.7234457    0.63860  11.7%     -    0s
#>      0     0    0.70893    0   20    0.72345    0.70893  2.01%     -    0s
#>      0     0    0.70915    0    6    0.72345    0.70915  1.98%     -    0s
#>      0     0    0.70915    0    6    0.72345    0.70915  1.98%     -    0s
#> H    0     0                       0.7178264    0.70920  1.20%     -    0s
#>      0     0    0.70921    0    3    0.71783    0.70921  1.20%     -    0s
#>      0     0    0.70921    0    5    0.71783    0.70921  1.20%     -    0s
#>      0     0    0.70924    0    7    0.71783    0.70924  1.20%     -    0s
#>      0     0    0.70993    0    6    0.71783    0.70993  1.10%     -    0s
#>      0     0    0.70997    0    9    0.71783    0.70997  1.09%     -    0s
#>      0     0    0.70997    0    9    0.71783    0.70997  1.09%     -    0s
#>      0     2    0.71000    0    9    0.71783    0.71000  1.09%     -    0s
#> H  150    20                       0.7169912    0.71221  0.67%   2.5    0s
#> 
#> Cutting planes:
#>   Cover: 13
#>   MIR: 2
#>   Flow cover: 1
#>   RLT: 1
#> 
#> Explored 200 nodes (841 simplex iterations) in 0.08 seconds (0.05 work units)
#> Thread count was 1 (of 8 available processors)
#> 
#> Solution count 1: 0.716991 
#> No other solutions better than 0.716991
#> 
#> Optimal solution found (tolerance 0.00e+00)
#> Best objective 7.169911970196e-01, best bound 7.169911970196e-01, gap 0.0000%
s$approach <- c("abs epsilon", "wtd goal", "Chebyshev", "ref point")

# print solutions
print(as.data.frame(s))
#>   solution  status     cost      obj1      obj2     obj3 A1_action A2_action
#> 1        1 OPTIMAL 884.8571 0.6701859 0.4297357 1.616604      TRUE     FALSE
#> 2        1 OPTIMAL 884.8571 0.6701859 0.4297357 1.616604      TRUE     FALSE
#> 3        1 OPTIMAL 969.5956 0.5989621 0.4297357 1.748802      TRUE      TRUE
#> 4        1 OPTIMAL 884.8571 0.6701859 0.4297357 1.616604      TRUE     FALSE
#>   A3_action A4_action A5_action A6_action A7_action A8_action A9_action
#> 1      TRUE     FALSE      TRUE      TRUE      TRUE     FALSE     FALSE
#> 2      TRUE     FALSE      TRUE      TRUE      TRUE     FALSE     FALSE
#> 3      TRUE      TRUE      TRUE     FALSE     FALSE      TRUE     FALSE
#> 4      TRUE     FALSE      TRUE      TRUE      TRUE     FALSE     FALSE
#>   A10_action A11_action A12_action A13_action A14_action A15_action B1_action
#> 1       TRUE       TRUE       TRUE      FALSE      FALSE       TRUE      TRUE
#> 2       TRUE       TRUE       TRUE      FALSE      FALSE       TRUE      TRUE
#> 3       TRUE       TRUE      FALSE       TRUE       TRUE      FALSE      TRUE
#> 4       TRUE       TRUE       TRUE      FALSE      FALSE       TRUE      TRUE
#>   B2_action B3_action F1_project F2_project F8_project baseline_project_obj1
#> 1      TRUE      TRUE       TRUE      FALSE      FALSE                  TRUE
#> 2      TRUE      TRUE       TRUE      FALSE      FALSE                  TRUE
#> 3      TRUE      TRUE      FALSE       TRUE      FALSE                  TRUE
#> 4      TRUE      TRUE       TRUE      FALSE      FALSE                  TRUE
#>   F3_project F4_project baseline_project_obj2 F5_project F6_project F7_project
#> 1      FALSE      FALSE                  TRUE      FALSE      FALSE      FALSE
#> 2      FALSE      FALSE                  TRUE      FALSE      FALSE      FALSE
#> 3      FALSE      FALSE                  TRUE      FALSE       TRUE      FALSE
#> 4      FALSE      FALSE                  TRUE      FALSE      FALSE      FALSE
#>   F9_project F10_project baseline_project_obj3        F1        F2        F8
#> 1       TRUE       FALSE                  TRUE 0.7104848 0.3990748 0.2802669
#> 2       TRUE       FALSE                  TRUE 0.7104848 0.3990748 0.2802669
#> 3      FALSE       FALSE                  TRUE 0.1391247 0.7582101 0.2802669
#> 4       TRUE       FALSE                  TRUE 0.7104848 0.3990748 0.2802669
#>          F3         F4       F5        F6        F7        F9        F10
#> 1 0.3814154 0.04832029 0.381568 0.1507689 0.3692636 0.6485493 0.06645404
#> 2 0.3814154 0.04832029 0.381568 0.1507689 0.3692636 0.6485493 0.06645404
#> 3 0.3814154 0.04832029 0.381568 0.5995262 0.3692636 0.3319902 0.06645404
#> 4 0.3814154 0.04832029 0.381568 0.1507689 0.3692636 0.6485493 0.06645404
#>      approach
#> 1 abs epsilon
#> 2    wtd goal
#> 3   Chebyshev
#> 4   ref point
# }