Project prioritization problem constraints

A constraint can be added to a project prioritization problem() to ensure that solutions exhibit a specific characteristic.

Details

The following constraints can be added to a project prioritization problem():

add_locked_in_constraints()

Add constraints to ensure that certain actions are prioritized for funding.

add_locked_out_constraints()

Add constraints to ensure that certain actions are not prioritized for funding.

See also

Examples

# load data
data(sim_projects, sim_features, sim_actions)

# build problem with maximum richness objective and $150 budget
p1 <- problem(sim_projects, sim_actions, sim_features,
             "name", "success", "name", "cost", "name") %>%
     add_max_richness_objective(budget = 150) %>%
     add_binary_decisions()

# print problem
print(p1)

#> Project Prioritization Problem
#>   actions          F1_action, F2_action, F3_action, ... (6 actions)
#>   projects         F1_project, F2_project, F3_project, ... (6 projects)
#>   features         F1, F2, F3, ... (5 features)
#>   action costs:    min: 0, max: 103.22583
#>   project success: min: 0.81379, max: 1
#>   objective:       Maximum richness objective [budget (150)]
#>   targets:         none
#>   weights:         default
#>   decisions        Binary decision 
#>   constraints:     <none>
#>   solver:          default

# build another problem, and lock in the third action
p2 <- p1 %>%
      add_locked_in_constraints(c(3))

# print problem
print(p2)

#> Project Prioritization Problem
#>   actions          F1_action, F2_action, F3_action, ... (6 actions)
#>   projects         F1_project, F2_project, F3_project, ... (6 projects)
#>   features         F1, F2, F3, ... (5 features)
#>   action costs:    min: 0, max: 103.22583
#>   project success: min: 0.81379, max: 1
#>   objective:       Maximum richness objective [budget (150)]
#>   targets:         none
#>   weights:         default
#>   decisions        Binary decision 
#>   constraints:     <Locked in actions [1 locked units]>
#>   solver:          default

# build another problem, and lock out the second action
p3 <- p1 %>%
      add_locked_out_constraints(c(2))

# print problem
print(p3)

#> Project Prioritization Problem
#>   actions          F1_action, F2_action, F3_action, ... (6 actions)
#>   projects         F1_project, F2_project, F3_project, ... (6 projects)
#>   features         F1, F2, F3, ... (5 features)
#>   action costs:    min: 0, max: 103.22583
#>   project success: min: 0.81379, max: 1
#>   objective:       Maximum richness objective [budget (150)]
#>   targets:         none
#>   weights:         default
#>   decisions        Binary decision 
#>   constraints:     <Manually locked actions [1 locked units]>
#>   solver:          default

# \dontrun{
# solve problems
s1 <- solve(p1)

#> Gurobi Optimizer version 9.1.1 build v9.1.1rc0 (linux64)
#> Thread count: 2 physical cores, 4 logical processors, using up to 1 threads
#> Optimize a model with 47 rows, 47 columns and 102 nonzeros
#> Model fingerprint: 0xd108d30d
#> Variable types: 0 continuous, 42 integer (42 binary)
#> Semi-Variable types: 5 continuous, 0 integer
#> Coefficient statistics:
#>   Matrix range     [9e-02, 1e+02]
#>   Objective range  [1e+00, 1e+00]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [1e+00, 2e+02]
#> Found heuristic solution: objective 1.4456093
#> Presolve removed 16 rows and 12 columns
#> Presolve time: 0.00s
#> Presolved: 31 rows, 35 columns, 65 nonzeros
#> Variable types: 0 continuous, 35 integer (35 binary)
#> Presolved: 31 rows, 35 columns, 65 nonzeros
#> 
#> 
#> Root relaxation: objective 1.680145e+00, 11 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#> *    0     0               0       1.6801450    1.68015  0.00%     -    0s
#> 
#> Explored 0 nodes (11 simplex iterations) in 0.00 seconds
#> Thread count was 1 (of 4 available processors)
#> 
#> Solution count 1: 1.68015 
#> 
#> Optimal solution found (tolerance 0.00e+00)
#> Best objective 1.680145013696e+00, best bound 1.680145013696e+00, gap 0.0000%

s2 <- solve(p2)

#> Gurobi Optimizer version 9.1.1 build v9.1.1rc0 (linux64)
#> Thread count: 2 physical cores, 4 logical processors, using up to 1 threads
#> Optimize a model with 47 rows, 47 columns and 102 nonzeros
#> Model fingerprint: 0x48d94c11
#> Variable types: 0 continuous, 42 integer (42 binary)
#> Semi-Variable types: 5 continuous, 0 integer
#> Coefficient statistics:
#>   Matrix range     [9e-02, 1e+02]
#>   Objective range  [1e+00, 1e+00]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [1e+00, 2e+02]
#> Found heuristic solution: objective 1.4456093
#> Presolve removed 42 rows and 41 columns
#> Presolve time: 0.00s
#> Presolved: 5 rows, 6 columns, 10 nonzeros
#> Variable types: 0 continuous, 6 integer (6 binary)
#> 
#> Explored 0 nodes (0 simplex iterations) in 0.00 seconds
#> Thread count was 1 (of 4 available processors)
#> 
#> Solution count 1: 1.44561 
#> 
#> Optimal solution found (tolerance 0.00e+00)
#> Best objective 1.445609277954e+00, best bound 1.445609277954e+00, gap 0.0000%

s3 <- solve(p3)

#> Gurobi Optimizer version 9.1.1 build v9.1.1rc0 (linux64)
#> Thread count: 2 physical cores, 4 logical processors, using up to 1 threads
#> Optimize a model with 47 rows, 47 columns and 102 nonzeros
#> Model fingerprint: 0x6fb0a017
#> Variable types: 0 continuous, 42 integer (42 binary)
#> Semi-Variable types: 5 continuous, 0 integer
#> Coefficient statistics:
#>   Matrix range     [9e-02, 1e+02]
#>   Objective range  [1e+00, 1e+00]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [1e+00, 2e+02]
#> Found heuristic solution: objective 1.4456093
#> Presolve removed 22 rows and 19 columns
#> Presolve time: 0.00s
#> Presolved: 25 rows, 28 columns, 52 nonzeros
#> Variable types: 0 continuous, 28 integer (28 binary)
#> Presolved: 25 rows, 28 columns, 52 nonzeros
#> 
#> 
#> Root relaxation: objective 1.575441e+00, 9 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#> *    0     0               0       1.5754408    1.57544  0.00%     -    0s
#> 
#> Explored 0 nodes (9 simplex iterations) in 0.00 seconds
#> Thread count was 1 (of 4 available processors)
#> 
#> Solution count 1: 1.57544 
#> 
#> Optimal solution found (tolerance 0.00e+00)
#> Best objective 1.575440809243e+00, best bound 1.575440809243e+00, gap 0.0000%

# print the actions selected for funding in each of the solutions
print(s1[, sim_actions$name])

#> # A tibble: 1 x 6
#>   F1_action F2_action F3_action F4_action F5_action baseline_action
#>       <dbl>     <dbl>     <dbl>     <dbl>     <dbl>           <dbl>
#> 1         0         1         0         0         0               1

print(s2[, sim_actions$name])

#> # A tibble: 1 x 6
#>   F1_action F2_action F3_action F4_action F5_action baseline_action
#>       <dbl>     <dbl>     <dbl>     <dbl>     <dbl>           <dbl>
#> 1         0         0         1         0         0               1

print(s3[, sim_actions$name])

#> # A tibble: 1 x 6
#>   F1_action F2_action F3_action F4_action F5_action baseline_action
#>       <dbl>     <dbl>     <dbl>     <dbl>     <dbl>           <dbl>
#> 1         1         0         0         0         0               1

# }