Matrix parameters

Create a parameter that represents a matrix object.

numeric_matrix_parameter(
  name,
  value,
  lower_limit = .Machine$double.xmin,
  upper_limit = .Machine$double.xmax,
  symmetric = FALSE
)

binary_matrix_parameter(name, value, symmetric = FALSE)

Arguments

name

character name of parameter.

value

matrix object.

lower_limit

numeric values denoting the minimum acceptable value in the matrix. Defaults to the smallest possible number on the system.

upper_limit

numeric values denoting the maximum acceptable value in the matrix. Defaults to the smallest possible number on the system.

symmetric

logical must the must be matrix be symmetric? Defaults to FALSE.

Value

MiscParameter object.

Examples

# create matrix
m <- matrix(runif(9), ncol = 3)
colnames(m) <- letters[1:3]
rownames(m) <- letters[1:3]

# create a numeric matrix parameter
p1 <- numeric_matrix_parameter("m", m)
print(p1) # print it
#> <environment: 0x55774e6c3d58>
#> attr(,"class")
#> [1] "MiscParameter" "MiscParameter" "Parameter"     "pproto"       
#> [5] "proto"         "environment"  
p1$get() # get value
#>            a         b         c
#> a 0.21314931 0.9447748 0.2882372
#> b 0.21031074 0.2449280 0.8753579
#> c 0.03952069 0.7811226 0.2957501
p1$id # get id
#> id: 93842016-7783-4d49-b430-17a7aba38a59
p1$validate(m[, -1]) # check if parameter can be updated
#> [1] FALSE
#> attr(,"msg")
#> [1] "ncol(m) not equal to ncol(value)"
p1$set(m + 1) # set parameter to new values
p1$print() # print it again
#> [1] "m"

# create a binary matrix parameter
m <- matrix(round(runif(9)), ncol = 3)
colnames(m) <- letters[1:3]
rownames(m) <- letters[1:3]

# create a binary matrix parameter
p2 <- binary_matrix_parameter("m", m)
print(p2) # print it
#> <environment: 0x55774ccb1dc8>
#> attr(,"class")
#> [1] "MiscParameter" "MiscParameter" "Parameter"     "pproto"       
#> [5] "proto"         "environment"  
p2$get() # get value
#>   a b c
#> a 1 1 0
#> b 1 1 0
#> c 1 0 1
p2$id # get id
#> id: ebed9e50-ce26-45af-8f91-960295411e1a
p2$validate(m[, -1]) # check if parameter can be updated
#> [1] FALSE
#> attr(,"msg")
#> [1] "ncol(m) not equal to ncol(value)"
p2$set(m + 1) # set parameter to new values
p2$print() # print it again
#> [1] "m"